analysis/software/requirements.txt
astropy==3.1.2
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
bokeh==1.2.0
Bottleneck==1.2.1
certifi==2019.6.16
Click==7.0
cloudpickle==1.1.1
colorama==0.4.1
corsikaio==0.2.0
cycler==0.10.0
cytoolz==0.9.0.1
dask==1.2.2
decorator==4.4.0
defusedxml==0.6.0
distributed==1.28.1
entrypoints==0.3
eventio==0.20.1
h5py==2.9.0
heapdict==1.0.0
iminuit==1.3.3
importlib-metadata==0.23
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
kiwisolver==1.1.0
llvmlite==0.28.0
locket==0.2.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mock==3.0.5
more-itertools==7.2.0
msgpack==0.6.1
numba==0.43.1
numexpr==2.6.9
numpy==1.16.4
olefile==0.46
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
partd==0.3.10
peewee==3.9.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.13.0
prometheus-client==0.7.0
prompt-toolkit==2.0.9
psutil==5.6.2
ptyprocess==0.6.0
py==1.8.0
pycrypto==2.6.1
pyfact==0.22.2
Pygments==2.4.2
pymongo==3.8.0
PyMySQL==0.9.3
pyparsing==2.4.0
pyrsistent==0.15.2
pytest==5.2.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.4
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
simple-crypt==4.1.7
six==1.12.0
sklearn-pandas==1.8.0
sklearn2pmml==0.44.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.3
tables==3.5.1
tblib==1.4.0
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.1
zict==0.1.4
zipp==0.6.0
